+++ completions/apfel.zsh
@@ -55,6 +55,7 @@ _apfel() {
     '-f[Attach file content to prompt]:file:_files' \
     '--file[Attach file content to prompt]:file:_files' \
     '--schema[Constrain output to a JSON Schema file]:file:_files' \
+    '--messages[One-shot multi-turn from OpenAI messages JSON]:file:_files' \
     '1:command:(completions)' \
     '*:file:_files'
 }